winapi callbacks back to x86 code
x86 is stepped a basic block at a time, but our native winapi implementations
are fully synchronous.  This means winapi can't call into x86, but it's required
in some cases.

To make this work, winapi functions that may callback do a careful dance where
we create and register a closure that manually maintains state across the function
being reentered.

This isn't yet complete -- IDirectDraw7::EnumDisplayModes needs to actually pass
a DDSURFACEDESC2 to the callback function, and there's some registration and
cleanup logic that needs to be figured out, but I think it's mostly right.