cache pointer to op function in instruction cache
When we decode a basic block (to fill the icache), look up the implementation
function we'll use to run each instruction and save it in the cache alongside
the decoded instruction.  This was worth about 10% faster runs on the zip test.

This saves a table lookup once per instruction invocation.
This technique is called a "loop based threaded interpreter" in
http://www.emulators.com/docs/nx25_nostradamus.htm , which is a great
resource that I saved a link to because I should do more of what it suggests.