properly handle 'repe cmpsb'
At the human-facing level there are three rep prefixes, rep/repe/repne.
But at the x86 level there are only two values, and the meaning of those
values depends on which instruction you're prefixing.  Because of confusion
around this we were interpreting 'repe cmpsb' as if it had the behavior of
'rep cmpsb', despite that not existing.

Fixes more of #10.