make async shims return Futures
This has the consequence that the logic for how to execute these Futures
gets moved to the emulator-specific code, which allows the different backends
to make different choices.

It breaks Unicorn and Rosetta, but I have convinced myself that Unicorn's
current approach is broken for other reasons, and Rosetta currently doesn't
build for other reasons, so I am fine with restoring those later.

From a patch from Luke Street <luke@street.dev>