use builtin dlls for shim resolution
Rather than poking magic numbers into the IAT, shims now are implemented
as real DLLs that make calls to retrowin32_syscall(), which is a function
that each emulation backend can expose.

For x86-emu this function calls sysenter and we hook that in emulation to
invoke shims.  For x86-64 this plumbs through to the CPU mode switching code.

A longer blog post about this change, with some diagrams:
  https://neugierig.org/software/blog/2024/09/retrowin32-syscalls.html