winapi: redo windows message queue
A MessageQueue type captures the details of WM_PAINT and WM_TIMER.

This fixes the bug where demos would sometimes freeze if you moused into
them, which was introduced in commit 30d1bb3e.  My root cause is that the
expected behaviors of the involved functions was unclear, so this clean
up not only fixes that but also tries to make the expectations clearer.